testing: build binaries and image with coverage support
For E2E testing we need binaries that emit a coverage profile. In Go,
only test binaries can do that, so we use the trick from
https://blog.cloudflare.com/go-coverage-with-external-tests/ and build
test binaries that execute the same code as normal binaries. The
coverage mode has to be chosen during compilation. We use "atomic"
because it gives us most information and works with the concurrent
goroutines in the driver.

"cmd" is now reserved for building binaries again. All real unit tests
are under "pkg", including the pmem-ns-init command line testing.

All three binaries have support for writing a termination message.