testing: create unique coverage files
When using -test.coverprofile, the name is fixed and the data from
older invocations get overwritten. That's inconvenient because testing
might include restarting pods.

We could deal with this in a shell wrapper, but then the difference
between production and testing deployments becomes bigger. An easier
solution is to add a new -coverprofile flag that, if set, chooses a
unique file name and then re-runs the same binary again with
-test.coverprofile set to that name.