clean up PROXY env handling
The top-level Makefile is a better place to define
them. clear-govm.make no longer had any of the code that it described,
only the proxy variable handling.

We supported upper and lower case env variable names for the cluster
setup, but Docker build arguments only supported lower case.

PROXY_ENV was only used once and therefore doesn't have to be a
separate variable.