support and document parallel clusters
All relevant settings are now documented in test-config.sh and set up
in such a way that existing env variables are used instead of the
default values from that file or the test-config.d files. This makes
it possible to override the values for individual clusters.

The registry handling is fully specified via test-config.sh,
additional insecure registries and the registry used for the PMEM-CSI
images can be configured separately. `make push-images` uses the same
settings instead of the previous "localhost:5000".

Redundant defaults for settings were removed as much as possible from
the scripts.

TEST_CLEAR_LINUX_VERSION is the new setting for the version
number (simpler than specifying a full image name).

TEST_CLEAR_LINUX_POST_INSTALL gets removed because it's unused and the
implementation no longer matched the description.