deploy: fix 1.14 and 1.15, webhook was missing
PR #530 broke deployment on 1.14 and 1.15 because it made all
components of a deployment mandatory but didn't add the "webhook"
link.

What further complicates it is that 1.14 does not support
objectSelector yet. For the sake of simplicity, the driver deployment
still happens in the "default" namespace on 1.14, but that namespace
is excluded from the webhook.