WIP: expose metric data from sidecar
The usual way of telling Prometheus to scrape a pod only works for a
single
endpoint (https://github.com/prometheus/prometheus/issues/3756).

This commit tries to solve that by merging data from the
external-provisioner sidecar into the data reported by the driver, but
that does not work because the data is partly the same (Go runtime)
and/or might become the same (CSI function call counts).